+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-device-breakpoints",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "Component for managing matchMedia queries in react.",
   "main": "lib/index.js",
   "scripts": {
