+++ src/consumer.js
@@ -1,0 +1,2 @@
+import Context from './context'
+export default Context.Consumer
