+++ src/index.js
@@ -1,57 +1,9 @@
-import React, { createContext, useEffect, useState, useContext } from 'react'
-import PropTypes from 'prop-types'
-
-const Breakpoints = createContext()
-
-const useBreakpoints = () => useContext(Breakpoints)
-
-const Media = Breakpoints.Consumer
-
-const breakpointChange = (updateState, bp) => ({ matches }) => {
-    updateState({
-      [bp.name]: matches
-    })
-}
-
-const BreakpointsProvider = ({ breakpoints, children }) => {
-    const [device, updateDevice] = useState(breakpoints.reduce((acc, bp) => ({
-        ...acc,
-        [bp.name]: bp.query.matches
-    }),{
-        isTouchDevice: /Mobi|Tablet|iPad|iPhone|Android/.test(window.navigator.userAgent)
-    }))
-    const updateState = (bp) => {
-        updateDevice({
-            ...device,
-            ...bp
-        })
-    }
-    useEffect(() => {
-        const listeners = breakpoints.map((bp, i) => {
-            const listener = breakpointChange(updateState, bp)
-            bp.query.addListener(listener)
-            return listener
-        })
-        return () => {
-          breakpoints.forEach((bp, i) => {
-            bp.query.removeListener(listeners[i])
-          })
-        }
-    }, [])
-    return <Breakpoints.Provider value={device}>{children}</Breakpoints.Provider>
-}
-
-BreakpointsProvider.propTypes = {
-    breakpoints: PropTypes.arrayOf(
-      PropTypes.shape({
-        name: PropTypes.string.isRequired,
-        query: PropTypes.object.isRequired
-      })
-    ).isRequired
-}
+import BreakpointsProvider from './provider'
+import Media from './consumer'
+import useBreakpoints from './useBreakpoints'
 
 export {
-    BreakpointsProvider,
-    useBreakpoints,
-    Media
+  BreakpointsProvider,
+  Media,
+  useBreakpoints
 }
