+++ src/useBreakpoints.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import Context from './context'
+
+const useBreakpoints = () => useContext(Context)
+
+export default useBreakpoints
